+++ source/rtcw_sp/src/botlib/be_aas_route.c
@@ -2030,7 +2030,7 @@ void AAS_DecompressVis( byte *in, int numareas, byte *decompressed ) {
 
 	//row = (numareas+7)>>3;
 	out = decompressed;
-	end = ( byte * )( (int)decompressed + numareas );
+	end = ( byte * )( (intptr_t)decompressed + numareas );
 
 	do
 	{
